renamed AztecRN package to ReactAztec to better match classes naming
