+++ android/Application/src/main/java/com/example/android/ReactAztec/ReactAztecView.java
@@ -1,4 +1,4 @@
-package com.example.android.AztecRN;
+package com.example.android.ReactAztec;
 
 import android.content.Context;
 import android.support.annotation.Nullable;
