+++ android/Application/src/main/java/com/example/android/common/activities/SampleRNBaseActivity.java
@@ -19,7 +19,7 @@ package com.example.android.common.activities;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 
-import com.example.android.AztecRN.ReactAztecPackage;
+import com.example.android.ReactAztec.ReactAztecPackage;
 import com.example.android.recyclerview.BuildConfig;
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.common.LifecycleState;
